% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguinWings.R
\docType{data}
\name{penguinWings}
\alias{penguinWings}
\alias{hum1}
\alias{hum2}
\alias{hum3}
\alias{rad1}
\alias{rad2}
\alias{rad3}
\alias{uln1}
\alias{uln2}
\alias{uln3}
\title{Penguin partial wing skeleton landmarks dataset}
\format{
A set of 'block' class objects.

An object of class \code{block} of length 1.

An object of class \code{block} of length 1.

An object of class \code{block} of length 1.

An object of class \code{block} of length 1.

An object of class \code{block} of length 1.

An object of class \code{block} of length 1.

An object of class \code{block} of length 1.

An object of class \code{block} of length 1.
}
\usage{
data(penguinWings)

hum2

hum3

rad1

rad2

rad3

uln1

uln2

uln3
}
\description{
The multi-part objects in this example are partial wing skeletons comprised of humerus, radius and ulna. These partial wing skeletons are from 15 extant species of penguin and five fossil species of penguin, which together constitute a dataset of 60 wing bones. Shape data from the multi-part objects are provided as landmark configurations. Three sets of landmarks were produced for each of the digital replicas so that the replicates could be averaged to mitigate effects of placement error. The nine sets of landmark configurations (i.e. three sets from each of the three bones) were separately read into \code{R} using \code{readPts} with \code{gpa = FALSE} (i.e. generalised Procrustes transformation not performed). Metadata for each bone are available on github (https://github.com/aharmer/morphoBlocks).
}
\keyword{datasets}
